Update bugfix stage2.py
Fix the ranking function and make it independent of a limiting magnitude that does not affect the ranking in the first place. Ranking is based on the background noise of a seeing FWHM and by that is close to the frozen package. Note that an image which is not the lowest FWHM can still be selected given the SNR tradeoff.